// apps/desktop/src/renderer/lib/types.ts
export interface FileAttachment {
	type: "file"
	url: string
	mediaType?: string
	filename?: string
}